pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qaas"
version = "0.1.0"
description = "Quantum-as-a-Service utilizes IQMProvider, IQMBackend, IQMJob and HEAppE to provider access to quantum machine"
authors = [
    { name="Jan Swiatkowski", email="jan.swiatkowski@vsb.cz" },
    { name="Jakub Konvička", email="jakub.konvicka@vsb.cz" },
    { name="Jan Martinovič", email="jan.martinovic@vsb.cz" },
    { name="Ladislav Foltyn", email="ladislav.foltyn@vsb.cz" }
]
readme = "README.rst"
requires-python = ">=3.11, <3.12"
license = { text = "MIT" }
dependencies = [
    "qiskit==1.2.4", # its just compatible
    "iqm-client[qiskit]==33.0.*",
    "iqm-exa-common==27.4.*",
    "iqm-station-control-client==12.0.*",
    "iqm-data-definitions==2.19",
    "iqm-pulla==12.0.*",
    "qiskit_aer>=0.15.0,<1.0.0",
    "Py4HEAppE>=2.5.1",
    "cryptography>=43.0.0",
    "bcrypt>=4.2.0",
    "cffi>=1.17.1",
    "click>=8.1.7",
    "jwcrypto==1.5.*",
    "PyJWT[crypto]==2.10.*",
    "numpy==2.3.3",
    "truststore"
]

#[project.urls]
#Homepage = ""

[tool.setuptools.packages.find]
where = ["."]

[tool.uv.sources]
python-http-irods-client = { git = "https://opencode.it4i.eu/lexis-platform/data/python-http-irods-client.git", rev = "1.1.4" }
